requirements.txt
pyautogen
openai
yfinance
matplotlib
langchain
flaml
pandas